Separate createSubscriptionWithExistingPaymentMethod with createSubscription
